returned to previous models_torch with custombatchnorm
